Change ProjectChangesetAfterFinalizedEvent to ProjectChangeRequestApprovedEvent
