+++ src/components/project-change-request/events/index.ts
@@ -1,1 +1,1 @@
-export * from './project-changeset-after-finalized.event';
+export * from './project-change-request-approved.event';
