+++ src/components/project-change-request/events/project-change-request-approved.event.ts
@@ -1,7 +1,7 @@
 import { Session, UnsecuredDto } from '../../../common';
 import { ProjectChangeRequest } from '../dto';
 
-export class ProjectChangesetAfterFinalizedEvent {
+export class ProjectChangeRequestApprovedEvent {
   constructor(
     readonly changeRequest: UnsecuredDto<ProjectChangeRequest>,
     readonly session: Session
